fix: norm error
