+++ src/include/ms_error.h
@@ -6,14 +6,12 @@
 /*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/03 07:10:06 by tookuyam          #+#    #+#             */
-/*   Updated: 2025/02/03 07:24:19 by tookuyam         ###   ########.fr       */
+/*   Updated: 2025/03/05 13:49:04 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MS_ERROR_H
 # define MS_ERROR_H
-
-# include "ms_error_type.h"
 
 void	ms_perror_cwd(const char *name);
