+++ src/include/ms_getopt.h
@@ -6,7 +6,7 @@
 /*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/31 09:58:22 by tookuyam          #+#    #+#             */
-/*   Updated: 2025/01/31 15:46:05 by tookuyam         ###   ########.fr       */
+/*   Updated: 2025/03/05 13:50:19 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include "ms_getopt_type.h"
 
 int		ms_getopt_init(
-			t_opting *opting, int argc, char *argv[],const char *optstring);
+			t_opting *opting, int argc, char *argv[], const char *optstring);
 bool	ms_getopt_parse(t_opting *opting);
 
 #endif
